+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] -- 0000-00-00
+
+### Added
+
+- Bugfix: Map text rendering not accounting for hi-res flag
+- Hook: TextRenderA
+- Hook: MapRenderB, MapRenderA
+
+### Fixed
+
+- QuickRace: Cantina not synced when setting number of racers
+- Updater: Mouse cursor not visible when showing restart message
+- Hook: TextRenderB insertion point after text render
+
 ## [0.1.3] -- 2024-04-30
 
 ### Added